+++ src/test/scala/behaviors/builders/BinaryTreeSpec.scala
@@ -2,7 +2,8 @@ package maze.behaviors
 
 import maze.behaviors.{ Linkage, Distance }
 import maze.behaviors.builders.{ BinaryTree, Generator }
-import maze.classes.{ Cell, Grid, Coordinates, MazeRequest, Algorithm }
+import maze.classes.{ SquareNeighbors, SquareCell, RectangleGrid, Coordinates, MazeRequest, Algorithm }
+import maze.classes.MazeType._
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers._
@@ -10,26 +11,26 @@ import org.scalatest.GivenWhenThen
 
 class BinaryTreeSpec extends AnyFlatSpec with GivenWhenThen {
 
-  case object module extends BinaryTree {
-    case object _linkage extends Linkage
-    override type LINKAGE = Linkage
+  case object module extends BinaryTree[SquareNeighbors, SquareCell, RectangleGrid] {
+    case object _linkage extends Linkage[SquareNeighbors, SquareCell, RectangleGrid]
+    override type LINKAGE = Linkage[SquareNeighbors, SquareCell, RectangleGrid]
     override val linker = _linkage
 
-    case object _distance extends Distance
-    override type DISTANCE = Distance
+    case object _distance extends Distance[SquareNeighbors, SquareCell, RectangleGrid]
+    override type DISTANCE = Distance[SquareNeighbors, SquareCell, RectangleGrid]
     override val distance = _distance
   }
 
   "BinaryTree" should "generate a 5x5 maze using Binary Tree and print it to screen" in {
     Given("5x5 grid")
-    val grid = Grid(5, 5, Coordinates(0, 4), Coordinates(4, 0))
+    val grid = RectangleGrid(5, 5, Coordinates(0, 4), Coordinates(4, 0))
     When("generating maze using BinaryTree")
-    val generated: Grid = module.generate(grid)
+    val generated = module.generate(grid)
     Then("generated maze should have height of 5 cells")
-    generated.rows should be (5)
+    generated.height should be (5)
     generated.cells.length should be (5)
     Then("generated maze should have width of 5 cells")
-    generated.columns should be (5)
+    generated.width should be (5)
     generated.cells.count(c => c.length == 5) should be (5)
     Then("resulting maze is a perfect maze")
     generated.isPerfectMaze() shouldBe (true) 
@@ -42,14 +43,14 @@ class BinaryTreeSpec extends AnyFlatSpec with GivenWhenThen {
 
   it should "generate a 20x20 maze and print to screen" in {
     Given("20x20 grid")
-    val grid = Grid(20, 20, Coordinates(0, 19), Coordinates(19, 0))
+    val grid = RectangleGrid(20, 20, Coordinates(0, 19), Coordinates(19, 0))
     When("generating maze using BinaryTree")
-    val generated: Grid = module.generate(grid)
+    val generated = module.generate(grid)
     Then("generated maze should have height of 20 cells")
-    generated.rows should be (20)
+    generated.height should be (20)
     generated.cells.length should be (20)
     Then("generated maze should have width of 20 cells")
-    generated.columns should be (20)
+    generated.width should be (20)
     generated.cells.count(c => c.length == 20) should be (20)
     Then("resulting maze is a perfect maze")
     generated.isPerfectMaze() shouldBe (true) 
@@ -61,14 +62,14 @@ class BinaryTreeSpec extends AnyFlatSpec with GivenWhenThen {
 
   it should "generate a 8x8 maze and print to screen" in {
     Given("8x8 grid")
-    val grid = Grid(8, 8, Coordinates(0, 7), Coordinates(7, 0))
+    val grid = RectangleGrid(8, 8, Coordinates(0, 7), Coordinates(7, 0))
     When("generating maze using BinaryTree")
-    val generated: Grid = module.generate(grid)
+    val generated = module.generate(grid)
     Then("generated maze should have height of 8 cells")
-    generated.rows should be (8)
+    generated.height should be (8)
     generated.cells.length should be (8)
     Then("generated maze should have width of 8 cells")
-    generated.columns should be (8)
+    generated.width should be (8)
     generated.cells.count(c => c.length == 8) should be (8)
     Then("resulting maze is a perfect maze")
     generated.isPerfectMaze() shouldBe (true) 
@@ -80,14 +81,14 @@ class BinaryTreeSpec extends AnyFlatSpec with GivenWhenThen {
 
   it should "generate a 10x10 maze and print to screen" in {
     Given("10x10 grid")
-    val grid = Grid(10, 10, Coordinates(0, 9), Coordinates(9, 0))
+    val grid = RectangleGrid(10, 10, Coordinates(0, 9), Coordinates(9, 0))
     When("generating maze using BinaryTree")
-    val generated: Grid = module.generate(grid)
+    val generated = module.generate(grid)
     Then("generated maze should have height of 10 cells")
-    generated.rows should be (10)
+    generated.height should be (10)
     generated.cells.length should be (10)
     Then("generated maze should have width of 10 cells")
-    generated.columns should be (10)
+    generated.width should be (10)
     generated.cells.count(c => c.length == 10) should be (10)
     Then("resulting maze is a perfect maze")
     generated.isPerfectMaze() shouldBe (true) 
@@ -99,14 +100,14 @@ class BinaryTreeSpec extends AnyFlatSpec with GivenWhenThen {
 
   it should "generate a 11x11 maze and print to screen" in {
     Given("11x11 grid")
-    val grid = Grid(11, 11, Coordinates(0, 10), Coordinates(10, 0))
+    val grid = RectangleGrid(11, 11, Coordinates(0, 10), Coordinates(10, 0))
     When("generating maze using BinaryTree")
-    val generated: Grid = module.generate(grid)
+    val generated = module.generate(grid)
     Then("generated maze should have height of 11 cells")
-    generated.rows should be (11)
+    generated.height should be (11)
     generated.cells.length should be (11)
     Then("generated maze should have width of 11 cells")
-    generated.columns should be (11)
+    generated.width should be (11)
     generated.cells.count(c => c.length == 11) should be (11)
     Then("resulting maze is a perfect maze")
     generated.isPerfectMaze() shouldBe (true) 
@@ -117,19 +118,19 @@ class BinaryTreeSpec extends AnyFlatSpec with GivenWhenThen {
   }
   
   it should "honor start and goal coordinates specified in MazeRequest when generating a large non-square grid" in {
-    case object module extends BinaryTree {
-      case object _linkage extends Linkage
-      override type LINKAGE = Linkage
+    case object module extends BinaryTree[SquareNeighbors, SquareCell, RectangleGrid] {
+      case object _linkage extends Linkage[SquareNeighbors, SquareCell, RectangleGrid]
+      override type LINKAGE = Linkage[SquareNeighbors, SquareCell, RectangleGrid]
       override val linker = _linkage
-      case object _distance extends Distance
-      override type DISTANCE = Distance
+      case object _distance extends Distance[SquareNeighbors, SquareCell, RectangleGrid]
+      override type DISTANCE = Distance[SquareNeighbors, SquareCell, RectangleGrid]
       override val distance = _distance
     }
     Given("JSON for a 52x29 BinaryTree request")
-    val json = """{"width":"52","height":"29","algorithm":"BinaryTree","startX":"14","startY":"0","goalX":"14","goalY":"28","mazeType":"Solved"}""" 
+    val json = """{"mazeType":"Square","width":"52","height":"29","algorithm":"BinaryTree","startX":"14","startY":"0","goalX":"14","goalY":"28","mazeType":"Solved"}""" 
     When("generating the grid")
     val request: MazeRequest = MazeRequest(json)
-    var grid: Grid = Generator.generate(request)
+    var grid = Generator.generate(request)
     Then("grid's start should be 14,0")
     grid.startCoords should equal (Coordinates(14, 0))
     info("START COORDS: " + grid.startCoords.toString())
