+++ src/test/scala/behaviors/builders/RecursiveBacktrackerSpec.scala
@@ -2,7 +2,8 @@ package maze.behaviors.builders
 
 import maze.behaviors.{ Linkage, Distance }
 import maze.behaviors.builders.RecursiveBacktracker
-import maze.classes.{ Cell, Grid, Coordinates, MazeRequest, Algorithm }
+import maze.classes.{ SquareNeighbors, SquareCell, RectangleGrid, Coordinates, MazeRequest, Algorithm }
+import maze.classes.MazeType._
 import maze.classes.Algorithm._
 
 import org.scalatest.flatspec.AnyFlatSpec
@@ -10,22 +11,22 @@ import org.scalatest.matchers.should.Matchers._
 import org.scalatest.GivenWhenThen
 
 class RecursiveBacktrackerSpec extends AnyFlatSpec with GivenWhenThen {
-  case object module extends RecursiveBacktracker {
-    case object _linkage extends Linkage
-    override type LINKAGE = Linkage
+  case object module extends RecursiveBacktracker[SquareNeighbors, SquareCell, RectangleGrid] {
+    case object _linkage extends Linkage[SquareNeighbors, SquareCell, RectangleGrid]
+    override type LINKAGE = Linkage[SquareNeighbors, SquareCell, RectangleGrid]
     override val linker = _linkage
 
-    case object _distance extends Distance
-    override type DISTANCE = Distance
+    case object _distance extends Distance[SquareNeighbors, SquareCell, RectangleGrid]
+    override type DISTANCE = Distance[SquareNeighbors, SquareCell, RectangleGrid]
     override val distance = _distance
   }
   
   "RecursiveBacktracker" should "generate a 6x6 maze using RecursiveBacktracker and print it to screen" in {
     Given("6x6 grid")
     val dim: Int = 6
-    val grid = Grid(dim, dim, Coordinates(0, dim - 1), Coordinates(dim - 1, 0))
+    val grid = RectangleGrid(dim, dim, Coordinates(0, dim - 1), Coordinates(dim - 1, 0))
     When("generating maze using RecursiveBacktracker algorithm")
-    val generated: Grid = module.generate(grid)
+    val generated = module.generate(grid)
     Then("resulting maze should contain no stranded unreachable cells") 
     generated.isFullyConnected() shouldBe (true)
     Then("resulting maze is a perfect maze")
@@ -37,9 +38,9 @@ class RecursiveBacktrackerSpec extends AnyFlatSpec with GivenWhenThen {
   it should "generate a 30x30 maze using RecursiveBacktracker and print it to screen" in {
     Given("30x30 grid")
     val dim: Int = 30
-    val grid = Grid(dim, dim, Coordinates(0, dim - 1), Coordinates(dim - 1, 0))
+    val grid = RectangleGrid(dim, dim, Coordinates(0, dim - 1), Coordinates(dim - 1, 0))
     When("generating maze using RecursiveBacktracker algorithm")
-    val generated: Grid = module.generate(grid)
+    val generated = module.generate(grid)
     Then("resulting maze should contain no stranded unreachable cells") 
     generated.isFullyConnected() shouldBe (true)
     Then("resulting maze is a perfect maze")
@@ -50,9 +51,9 @@ class RecursiveBacktrackerSpec extends AnyFlatSpec with GivenWhenThen {
 
   it should "honor start and goal coordinates specified in MazeRequest when generating a non-square RecursiveBacktracker maze grid" in {
     Given("5x10 RecursiveBacktracker request")
-    val request = MazeRequest(5, 10, Algorithm.RecursiveBacktracker, Coordinates(0, 9), Coordinates(4, 0))
+    val request = MazeRequest(Square, 5, 10, Algorithm.RecursiveBacktracker, Coordinates(0, 9), Coordinates(4, 0))
     When("generating the grid")
-    var grid: Grid = Generator.generate(request)
+    var grid = Generator.generate(request).asInstanceOf[RectangleGrid]
     Then("grid's start should be southwest")
     grid.startCoords should equal (Coordinates(0, request.height - 1))
     info("START COORDS: " + grid.startCoords.toString())
